Move updateMetadataSource and verifyMetadataSource to DSL.hs
